+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 ~~~~~~~~~
 
+0.11.0 (May 2024)
+-----------------
+
+* Updated pytket version requirement to 1.28.
+* add support for BitWiseOp.ONE, BitWiseOp.ZERO in conversion
+
 0.10.1 (April 2024)
 -------------------